// dashboard-ts/src/api-types.d.ts
export type AllOfficersOfBandobasType = Array<{
	id: number;
	police_user: {
		id: number;
		name: string;
		phone: string;
		rank: string;
		department: string;
		image_url: string;
	};
	weapons: string;
	fcm_token: string;
	device_id: string;
	duty_start_time: string;
	duty_end_time: string;
	created_by: number;
	assigned_nfc_device: number;
}>;